pyproject.toml
[project]
name = "piu-piu"
version = "0.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.12.3",
    "pydantic-yaml>=1.6.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pylint>=4.0.2",
    "pyright>=1.1.407",
    "ruff>=0.14.2",
]

[tool.pyright]
strict = ["**"]
